Added new export files for print testing
Printing out ribs and spars, along with aileron for testing. I know I want to throw something on the ends of the aircraft to seal the skins in. I am thinking something like a rounded cap that also connects to the edge spars. This way it helps secure those edges on full.

I also must start working on the development of the servo mounts now. I am going to probably create a new component inside the wings file for the control horn, based on already done measurements, then attach it to determine how my servo arm will fit.