Adjustments to Files
We have removed previous gcode test files, and have completely added a full list of the first prototype 3D models.

Added to the resource tracking excel sheet with the expected weights of all parts and 3D printed components with the exception of the landing gear.

Filled in status notes in the overall documentation file.